third_party/WebKit/LayoutTests/fast/css/getComputedStyle/computed-style-recalc-expected.txt

Check that getComputedStyle causes a style recalc when necessary, and only when necessary.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS internals.updateStyleAndReturnAffectedElementCount() is 0
PASS internals.updateStyleAndReturnAffectedElementCount() is 2
PASS internals.updateStyleAndReturnAffectedElementCount() is 2
PASS successfullyParsed is true

TEST COMPLETE
